Add Bazel build rules.
This adds support for building effcee with Bazel (https://bazel.build).

It automatically pulls the googletest and re2 repositories when building
with Bazel.  I did not add a sha256 attribute to fix the version of the
external repos to get, so by default this tracks top-of-master for those
repos.

It's easy to fix the versions by adding the sha256 attribute that Bazel
suggests when it starts.