Remove iterator diff from two different StringPieces
The behavious is technically undefined when you take the diff of
iterators from two different string views. This is why MSVC will cause
an assert in their debug builds.

I modified that code to find the captured string in the input to be able
to calculate the length of the prefixe that needs to be removed. It will
be much slower, but it avoids the assert.

Fixes #52