[Kokoro] Use merge-base for format check
When doing the clang format check, the current branch is compared with
main to see which files changed, and if there are any formating issues
in those files. This can cause false positives if the current branch is
too far behind main. If another PR fixed the formating in main, then
that will be flagged as an error because it looks like the current PR
will undo that change.

The fix is to use the merge-base for main and HEAD to find the first
common ancestor of the two to determine which files are changed in the
current branch.

Bug: crbug.com/414577655